I did myself a monkeypatch ¯\_(ツ)_/¯
Once I realized that the GCal API returns colors as part of the event
payload, I wanted to use that as a condition for turning on my light,
but that meant the list of conditions was getting a bit long.

So, I decided to just add a valid? method on the GCal Event class itself
since that seemed a bit cleaner than having my Calendar object do
calculations about the Event's properties.