During install, generate a script to run with cron
The previous script that cron would run was very specific to my local
machine and would require some heavy changes to the file in order to
work on other computers.

Now, the bin/install script generates a script for cron to run that's
totally dependent on the user's environment, making this program more
flexible. Still requires RVM though.